cargo update

+++ rain-cli/src/main.rs
@@ -13,7 +13,7 @@ fn main() -> ExitCode {
 fn fallible_main() -> Result<(), ()> {
     let cli = Cli::parse();
     match cli.command {
-        RainCommand::Run { script } => {
+        RainCommand::RunScript { script } => {
             let v = rain_lang::run_stderr(script, rain_lang::config::Config::default())?;
             eprintln!("{v:?}");
             Ok(())
@@ -21,7 +21,18 @@ fn fallible_main() -> Result<(), ()> {
         RainCommand::Clean => {
             let config = rain_lang::config::Config::default();
             let clean_path = &config.base_cache_dir;
-            eprintln!("Removing {}", clean_path.display());
+            eprintln!("removing {}", clean_path.display());
+            let metadata = match std::fs::metadata(clean_path) {
+                Ok(metadata) => metadata,
+                Err(err) => {
+                    eprintln!("could not stat cache directory: {err}");
+                    return Err(());
+                }
+            };
+            if !metadata.is_dir() {
+                eprintln!("failed {} is not a directory", clean_path.display());
+                return Err(());
+            }
             if let Err(err) = std::fs::remove_dir_all(clean_path) {
                 eprintln!("clean failed: {err}");
                 return Err(());
@@ -40,6 +51,6 @@ struct Cli {
 
 #[derive(Debug, Subcommand)]
 pub enum RainCommand {
-    Run { script: PathBuf },
+    RunScript { script: PathBuf },
     Clean,
 }
